Add Semantic Versioning Library to Common
