+++ ChairManager/BinaryVersionCheck.cpp
@@ -14,23 +14,24 @@
 #include <curlpp/Info.hpp>
 #include <curlpp/Infos.hpp>
 #include "UpdateURL.h"
+#include "../Common/Chairloader/SemanticVersion.h"
 
 static std::string m_latestVersion;
 
 
-VersionCheck::DLLVersion VersionCheck::getInstalledChairloaderVersion() {
+SemanticVersion VersionCheck::getInstalledChairloaderVersion() {
     auto szVersionFile = ChairManager::Get().GetGamePath() / PathUtils::GAME_BIN_DIR / "Chairloader.dll";
     return getBinaryVersion(szVersionFile);
 }
 
-VersionCheck::DLLVersion VersionCheck::getPackagedChairloaderVersion() {
+SemanticVersion VersionCheck::getPackagedChairloaderVersion() {
     auto szVersionFile = fs::path("Release") / "Chairloader.dll";
     return getBinaryVersion(szVersionFile);
 }
 
-VersionCheck::DLLVersion VersionCheck::getBinaryVersion(fs::path szVersionFile) {
+SemanticVersion VersionCheck::getBinaryVersion(fs::path szVersionFile) {
     auto strVersion = getBinaryVersionString(szVersionFile);
-    return DLLVersion(strVersion);
+    return SemanticVersion(strVersion);
 }
 
 std::string VersionCheck::getBinaryVersionString(fs::path szVersionFile) {
@@ -65,8 +66,8 @@ std::string VersionCheck::getLatestChairloaderVersionString() {
     return m_latestVersion;
 }
 
-VersionCheck::DLLVersion VersionCheck::getLatestChairloaderVersion() {
-    return DLLVersion(getLatestChairloaderVersionString());
+SemanticVersion VersionCheck::getLatestChairloaderVersion() {
+    return SemanticVersion(getLatestChairloaderVersionString());
 }
 
 void VersionCheck::fetchLatestVersion(bool bForce) {
@@ -111,36 +112,3 @@ void VersionCheck::fetchLatestVersion(bool bForce) {
     ChairManager::Get().setCachedLatestVersion(tag_name);
 }
 
-
-std::string VersionCheck::DLLVersion::String() {
-    std::string baseVersion = boost::str(boost::format("%d.%d.%d") % m_Major % m_Minor % m_Patch);
-    if(!m_ReleaseType.empty()){
-        baseVersion += "-" + m_ReleaseType;
-    }
-    if(m_Build != 0){
-        baseVersion += "+" + std::to_string(m_Build);
-    }
-    return baseVersion;
-}
-
-VersionCheck::DLLVersion::DLLVersion(std::string versionStr) {
-    // parse the semantic version using regex
-    std::regex versionRegex(R"(^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)(?:-((?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+([0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$)");
-    std::smatch match;
-    if (std::regex_match(versionStr, match, versionRegex)) {
-        m_Major = std::stoi(match[1]);
-        m_Minor = std::stoi(match[2]);
-        m_Patch = std::stoi(match[3]);
-        m_ReleaseType = match[4];
-        try{
-            m_Build = std::stoi(match[5]);
-        } catch (std::invalid_argument& e) {
-            m_Build = 0;
-        } catch (std::out_of_range& e) {
-            m_Build = 0;
-        } catch (std::exception& e) {
-            m_Build = 0;
-        }
-    }
-    m_VersionStr = versionStr;
-}
