+++ ChairManager/ChairInstallWizard.cpp
@@ -3,6 +3,7 @@
 #include "PathUtils.h"
 #include "ChairManager.h"
 #include "BinaryVersionCheck.h"
+#include "../Common/Chairloader/SemanticVersion.h"
 
 static const ImVec2 DEFAULT_WINDOW_SIZE = { 600, 400 };
 static ImVec2 WINDOW_SIZE = DEFAULT_WINDOW_SIZE;
